+++ packages/error-parser/src/reporters/console.ts
@@ -1,0 +1,5 @@
+import { LiqualityError } from '../LiqualityErrors';
+
+export function reportToConsole(error: LiqualityError) {
+  console.log(error);
+}