+++ packages/wallet-core/src/nameResolvers/nameResolver.ts
@@ -1,8 +1,8 @@
 import { Nullable } from '@chainify/types';
-import { ChainId } from '@liquality/cryptoassets';
+import { IAsset } from '@liquality/cryptoassets';
 
 export interface NameResolver {
   reverseLookup(address: string): Promise<Nullable<string>>;
-  lookupDomain(address: string, chainId: ChainId): Promise<Nullable<string>>;
+  lookupDomain(address: string, asset: IAsset): Promise<Nullable<string>>;
   isValidTLD(domain: string): Promise<boolean>;
 }