+++ packages/wallet-core/src/store/actions/enableDisableAssets.test.ts
@@ -4,8 +4,8 @@ import { Network } from '../types';
 
 let wallet: any;
 describe('disable and enable assets', () => {
+  jest.setTimeout(20000);
   beforeEach(async () => {
-    jest.useFakeTimers();
     wallet = await setupWallet(defaultWalletOptions);
     await wallet.dispatch.createWallet({
       key: '0x1234567890123456789012345678901234567890',