+++ packages/wallet-core/src/store/actions/sendNFTTransaction.ts
@@ -12,8 +12,7 @@ export const sendNFTTransaction = async (
   const { getters, commit, dispatch } = rootActionContext(context);
   const account: Account = getters.accountItem(accountId)!;
   const asset = getChain(network, account.chain).nativeAsset[0].code;
-  const client = getters.client({ network, walletId, chainId: account.chain });
-
+  const client = getters.client({ network, walletId, chainId: account.chain, accountId });
   const tx = await client.nft.transfer(nft.asset_contract!.address!, receiver, [nft.token_id!], values);
 
   const transaction: NFTSendHistoryItem = {