+++ packages/cryptoassets/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @liquality/cryptoassets
 
+## 1.14.11
+
+### Patch Changes
+
+- 5e73c63: feat: bump all versions from master
+
 ## 1.14.10
 
 ### Patch Changes
@@ -10,18 +16,20 @@
 
 ### Patch Changes
 
-- c9f3735: - fix address validation for evm chains
+- a45924d: - fix address validation for evm chains
 
 ## 1.14.7
 
 ### Patch Changes
 
+- 376246d: test: testing bump versions with beta tag
 - 7179204: chore: bump versions with correct script
 
 ## 1.14.6
 
 ### Patch Changes
 
+- f3bb88f: fix: uns name resolving
 - 9e0c685: core: bump both packages versions
 
 ## 1.14.5